work/airflow_aws/requirements.txt
# Apache Airflow依赖
apache-airflow>=2.7.0
apache-airflow-providers-amazon>=8.0.0

# SSH连接
paramiko>=3.3.1

# YAML解析
PyYAML>=6.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# 其他工具
python-dotenv>=1.0.0

